Move plotting module to root package
