+++ qupulse/__init__.pyi
@@ -8,6 +8,7 @@ from . import comparable
 from . import expressions
 from . import parameter_scope
 from . import serialization
+from . import plotting
 
 from .utils.types import MeasurementWindow, ChannelID
 
@@ -21,4 +22,5 @@ __all__ = ['pulses',
            'serialization',
            'MeasurementWindow',
            'ChannelID',
+           'plotting',
            ]
